Update rollup-plugin-ts, rearrange source so that .d.ts output works.
