+++ package.json
@@ -31,7 +31,7 @@
     "rollup-plugin-jscc": "^1.0.0",
     "rollup-plugin-node-resolve": "^4.0.1",
     "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-ts": "^1.2.14",
+    "rollup-plugin-ts": "^1.2.23",
     "semver": "^5.3.0",
     "terser": "^3.16.1",
     "tslib": "^1.9.3",
@@ -55,8 +55,8 @@
     "prebuild": "npm run lint",
     "build": "tsc && rollup -c && npm run build:min && npm run build:dts",
     "build:min": "npm run build:min:bundle && npm run build:min:legacy",
-    "build:min:bundle":"terser dist/pixi-animate.js -c -m -o dist/pixi-animate.min.js --source-map \"content='dist/pixi-animate.js.map',includeSources=true,filename='dist/pixi-animate.min.js.map',url='pixi-animate.min.js.map'\" --comments \"/pixi-animate.js - /\"",
-    "build:min:legacy":"terser dist/pixi-animate-legacy.js -c -m -o dist/pixi-animate-legacy.min.js --source-map \"content='dist/pixi-animate-legacy.js.map',includeSources=true,filename='dist/pixi-animate-legacy.min.js.map',url='pixi-animate-legacy.min.js.map'\" --comments \"/pixi-animate-legacy.js - /\"",
+    "build:min:bundle": "terser dist/pixi-animate.js -c -m -o dist/pixi-animate.min.js --source-map \"content='dist/pixi-animate.js.map',includeSources=true,filename='dist/pixi-animate.min.js.map',url='pixi-animate.min.js.map'\" --comments \"/pixi-animate.js - /\"",
+    "build:min:legacy": "terser dist/pixi-animate-legacy.js -c -m -o dist/pixi-animate-legacy.min.js --source-map \"content='dist/pixi-animate-legacy.js.map',includeSources=true,filename='dist/pixi-animate-legacy.min.js.map',url='pixi-animate-legacy.min.js.map'\" --comments \"/pixi-animate-legacy.js - /\"",
     "build:dts": "node scripts/buildAmbientTypes.js",
     "test": "floss --path tests/index.js",
     "coverage": "nyc npm test",
