+++ src/animate/DisplayObject.ts
@@ -1,8 +1,3 @@
-export * from './Container';
-export * from './Sprite';
-export * from './Graphics';
-export * from './Text';
-
 import { AnimateContainer } from './Container';
 import { AnimateSprite } from './Sprite';
 import { AnimateGraphics } from './Graphics';
