+++ src/animate/index.ts
@@ -1,12 +1,14 @@
-export { load } from './load';
-export { sound } from './sound';
-// eslint-disable-next-line @typescript-eslint/camelcase
-export { utils_ns as utils } from './utils';
-export { MovieClip } from './MovieClip';
-export { Scene } from './Scene';
-export { Timeline } from './Timeline';
-export { Tween } from './Tween';
-export { Animator } from './Animator';
-export { AnimatorTimeline } from './AnimatorTimeline';
-
-export const VERSION = '__VERSION__';
+export * from './load';
+export * from './sound';
+export * from './utils';
+export * from './MovieClip';
+export * from './Scene';
+export * from './Timeline';
+export * from './Tween';
+export * from './Animator';
+export * from './AnimatorTimeline';
+export * from './Container';
+export * from './Sprite';
+export * from './Graphics';
+export * from './Text';
+export * from './DisplayObject';
