+++ src/index.ts
@@ -1,4 +1,13 @@
-// Because AnimateAsset wants to import the entire library for typing purposes, they are separated
-// at this level to try to reduce odd references
-export * from './main';
+export { load, sound, utils, MovieClip, Scene, Timeline, Tween, Animator, AnimatorTimeline, DrawCommands } from './animate';
+
+export {
+    AnimateContainer as Container,
+    AnimateSprite as Sprite,
+    AnimateGraphics as Graphics,
+    AnimateText as Text,
+    AnimateDisplayObject as DisplayObject,
+} from './animate';
+
+export const VERSION = '__VERSION__';
+
 export * from './AnimateAsset';
