--- src/main.ts
@@ -1,10 +1,0 @@
-// For why this is separate, see index.ts
-export * from './animate';
-export {
-    AnimateDisplayObject as DisplayObject,
-    AnimateContainer as Container,
-    AnimateText as Text,
-    AnimateGraphics as Graphics,
-    AnimateSprite as Sprite,
-    DrawCommands,
-} from './mixins';
